# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/platforms": {
      "get": {
        "operationId": "PlatformController_findAll",
        "parameters": [],
        "responses": {
          "200": { "description": "" }
        },
        "tags": ["Platforms"]
      },
      "post": {
        "operationId": "PlatformController_create",
        "parameters": [],
        "responses": {
          "201": { "description": "" }
        },
        "tags": ["Platforms"]
      }
    },
    "/platforms/{id}": {
      "get": {
        "operationId": "PlatformController_findOne",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": ""}},
        "tags": ["Platforms"]
      },
      "patch": {
        "operationId": "PlatformController_update",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": ""}},
        "tags": ["Platforms"]
      },
      "delete": {
        "operationId": "PlatformController_remove",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": ""}},
        "tags": ["Platforms"]
      }
    }
  },
  "info": {
    "title": "Gamearr API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}